requirements.txt
pytorch
numpy
tqdm
torchsummary
decord
librosa
scipy
pystoi
pesq
soundfile
torchvision
torchaudio
transformers
diffusers
timm